move autOptions to a separate file
