+++ app/api/auth/[...nextauth]/route.ts
@@ -1,33 +1,6 @@
-import { mergeCarts } from "@/lib/db/cart";
-import { prisma } from "@/lib/db/prisma";
-import { env } from "@/lib/env";
-import { PrismaAdapter } from "@auth/prisma-adapter";
-import { NextAuthOptions } from "next-auth";
-import { Adapter } from "next-auth/adapters";
-import GoogleProvider from "next-auth/providers/google";
+import { authOptions } from "@/lib/auth";
 import NextAuth from "next-auth";
-
-const authOptions: NextAuthOptions = {
-  adapter: PrismaAdapter(prisma) as Adapter,
-  providers: [
-    GoogleProvider({
-      clientId: env.GOOGLE_CLIENT_ID,
-      clientSecret: env.GOOGLE_CLIENT_SECRET,
-    }),
-  ],
-  callbacks: {
-    session({ session, user }) {
-      session.user.id = user.id;
-      return session;
-    },
-  },
-  events: {
-    async signIn({ user }) {
-      await mergeCarts(user.id);
-    },
-  },
-};
 
 const handler = NextAuth(authOptions);
 
-export { handler as GET, handler as POST, authOptions };
+export { handler as GET, handler as POST };
