+++ lib/db/cart.ts
@@ -2,7 +2,7 @@ import { cookies } from "next/dist/client/components/headers";
 import { prisma } from "./prisma";
 import { Cart, CartItem, Prisma } from "@prisma/client";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "../auth";
 
 // ??
 export type CartWithProducts = Prisma.CartGetPayload<{
